+++ src/com/happyheng/service/SportRecordService.java
@@ -1,5 +1,5 @@
 package com.happyheng.service;
 
 public interface SportRecordService {
-	public int record(int sportId, float posx, float posy, String location);
+	public int record(int sportId, double posx, double posy, String location);
 }
